# api/resources/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/processes": {
      "post": {
        "operationId": "ProcessesController_create",
        "summary": "Criar processo avaliativo (US-08)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProcessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProcessesController_findAll",
        "summary": "Listar meus processos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/processes/{id}": {
      "get": {
        "operationId": "ProcessesController_findOne",
        "summary": "Detalhar processo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Processo não encontrado"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProcessesController_update",
        "summary": "Editar processo (apenas rascunho)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProcessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Processo não está em rascunho"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/processes/{id}/questions": {
      "post": {
        "operationId": "ProcessesController_addQuestions",
        "summary": "Selecionar perguntas do processo (US-09)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddQuestionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Perguntas adicionadas"
          },
          "400": {
            "description": "Pergunta inválida ou arquivada"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProcessesController_listQuestions",
        "summary": "Perguntas já vinculadas ao processo (US-09)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Perguntas do processo, em ordem"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/processes/{id}/candidates": {
      "post": {
        "operationId": "ProcessesController_addCandidates",
        "summary": "Adicionar candidatos ao processo (US-10)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCandidatesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Candidatos adicionados"
          },
          "400": {
            "description": "Candidato inválido ou duplicado"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProcessesController_listCandidates",
        "summary": "Candidatos já vinculados ao processo (US-10)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidatos do processo"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/processes/{id}/publish": {
      "patch": {
        "operationId": "ProcessesController_publish",
        "summary": "Publicar processo (US-11)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Processo sem perguntas ou sem candidatos"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/processes/{id}/close": {
      "patch": {
        "operationId": "ProcessesController_close",
        "summary": "Encerrar processo manualmente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/processes/{id}/candidates/panel": {
      "get": {
        "operationId": "ProcessesController_candidatePanel",
        "summary": "Painel de candidatos do processo (US-31)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidatePanelDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Processo não encontrado"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/processes/{id}/stats": {
      "get": {
        "operationId": "ProcessesController_stats",
        "summary": "Tempo de resposta e média do processo (US-32)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessStatsDto"
                }
              }
            }
          },
          "404": {
            "description": "Processo não encontrado"
          }
        },
        "tags": [
          "Processos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_listar",
        "summary": "Listar meus avisos (US-33/US-34) — consumir via polling",
        "parameters": [
          {
            "name": "naoLidas",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Notificações"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_marcarComoLida",
        "summary": "Marcar aviso como lido",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Notificação não encontrada"
          }
        },
        "tags": [
          "Notificações"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "API — Software de Processo Avaliativo",
    "description": "API REST para condução de processos avaliativos (avaliador e avaliado). Autenticação via JWT no cabeçalho Authorization: Bearer <token>.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProcessDto": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string",
            "example": "Processo Seletivo QA Trainee 2026"
          },
          "descricao": {
            "type": "string",
            "example": "Avaliação de fundamentos de QA e API REST."
          },
          "tempoMinutos": {
            "type": "number",
            "example": 90,
            "description": "Tempo limite em minutos"
          },
          "notaMinima": {
            "type": "number",
            "example": 70,
            "description": "Nota mínima de aprovação em %"
          },
          "modoAprovacao": {
            "type": "string",
            "enum": [
              "AUTOMATICO",
              "MANUAL",
              "AMBOS"
            ],
            "example": "AUTOMATICO"
          },
          "exibirGabarito": {
            "type": "boolean",
            "example": false,
            "description": "Exibir gabarito ao candidato após correção"
          },
          "dataInicio": {
            "type": "string",
            "example": "2026-07-01T08:00:00.000Z"
          },
          "dataFim": {
            "type": "string",
            "example": "2026-07-10T23:59:59.000Z"
          }
        },
        "required": [
          "titulo",
          "tempoMinutos",
          "notaMinima",
          "modoAprovacao"
        ]
      },
      "ProcessResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "object",
            "nullable": true
          },
          "tempoMinutos": {
            "type": "number"
          },
          "notaMinima": {
            "type": "number",
            "description": "Nota mínima de aprovação em %"
          },
          "modoAprovacao": {
            "type": "string",
            "enum": [
              "AUTOMATICO",
              "MANUAL",
              "AMBOS"
            ]
          },
          "exibirGabarito": {
            "type": "boolean"
          },
          "dataInicio": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "dataFim": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "RASCUNHO",
              "ATIVO",
              "ENCERRADO"
            ]
          },
          "totalPerguntas": {
            "type": "number"
          },
          "totalCandidatos": {
            "type": "number"
          },
          "criadoEm": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "titulo",
          "tempoMinutos",
          "notaMinima",
          "modoAprovacao",
          "exibirGabarito",
          "status",
          "totalPerguntas",
          "totalCandidatos",
          "criadoEm"
        ]
      },
      "UpdateProcessDto": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string",
            "example": "Processo Seletivo QA Trainee 2026"
          },
          "descricao": {
            "type": "string",
            "example": "Avaliação de fundamentos de QA e API REST."
          },
          "tempoMinutos": {
            "type": "number",
            "example": 90,
            "description": "Tempo limite em minutos"
          },
          "notaMinima": {
            "type": "number",
            "example": 70,
            "description": "Nota mínima de aprovação em %"
          },
          "modoAprovacao": {
            "type": "string",
            "enum": [
              "AUTOMATICO",
              "MANUAL",
              "AMBOS"
            ],
            "example": "AUTOMATICO"
          },
          "exibirGabarito": {
            "type": "boolean",
            "example": false,
            "description": "Exibir gabarito ao candidato após correção"
          },
          "dataInicio": {
            "type": "string",
            "example": "2026-07-01T08:00:00.000Z"
          },
          "dataFim": {
            "type": "string",
            "example": "2026-07-10T23:59:59.000Z"
          }
        }
      },
      "ProcessoPerguntaItemDto": {
        "type": "object",
        "properties": {
          "perguntaId": {
            "type": "string",
            "format": "uuid"
          },
          "ordem": {
            "type": "number",
            "example": 1,
            "description": "Ordem de exibição"
          }
        },
        "required": [
          "perguntaId",
          "ordem"
        ]
      },
      "AddQuestionsDto": {
        "type": "object",
        "properties": {
          "perguntas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessoPerguntaItemDto"
            }
          }
        },
        "required": [
          "perguntas"
        ]
      },
      "AddCandidatesDto": {
        "type": "object",
        "properties": {
          "candidatoIds": {
            "description": "IDs dos avaliados",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "candidatoIds"
        ]
      },
      "CandidatePanelDto": {
        "type": "object",
        "properties": {
          "tentativaId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "Nulo se o candidato ainda não iniciou"
          },
          "candidatoId": {
            "type": "string",
            "format": "uuid"
          },
          "candidatoNome": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Status da tentativa, ou NAO_INICIADA se o candidato ainda não começou",
            "enum": [
              "EM_ANDAMENTO",
              "ENVIADA",
              "CORRIGIDA",
              "NAO_INICIADA"
            ]
          },
          "scoreParcial": {
            "type": "object",
            "nullable": true
          },
          "scoreTotal": {
            "type": "object",
            "nullable": true
          },
          "aprovado": {
            "type": "object",
            "nullable": true
          },
          "abertasPendentes": {
            "type": "number"
          },
          "totalSaidas": {
            "type": "number",
            "description": "Total de saídas de janela (anti-cola)"
          },
          "tempoGastoSegundos": {
            "type": "object",
            "nullable": true,
            "description": "Tempo gasto em segundos"
          }
        },
        "required": [
          "tentativaId",
          "candidatoId",
          "candidatoNome",
          "status",
          "abertasPendentes",
          "totalSaidas"
        ]
      },
      "ProcessStatsDto": {
        "type": "object",
        "properties": {
          "totalCandidatos": {
            "type": "number"
          },
          "finalizados": {
            "type": "number"
          },
          "mediaTempoSegundos": {
            "type": "object",
            "nullable": true,
            "description": "Média de tempo em segundos"
          }
        },
        "required": [
          "totalCandidatos",
          "finalizados"
        ]
      },
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "NOVO_ENVIO",
              "RESULTADO_DISPONIVEL"
            ]
          },
          "mensagem": {
            "type": "string"
          },
          "link": {
            "type": "object",
            "nullable": true
          },
          "lida": {
            "type": "boolean"
          },
          "criadoEm": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tipo",
          "mensagem",
          "lida",
          "criadoEm"
        ]
      },
...
    }
  }
}